selerio/emergency-ai/requirements.txt
google-generativeai
geopy
requests
python-dotenv
aiohttp
asyncio
typing-extensions
adk
SpeechRecognition